opl dosbox extracted Timer
