+++ sdl2-sonic-drivers/CMakeLists.txt
@@ -65,6 +65,7 @@ install(FILES
     ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/Config.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/OplType.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/OplEmulator.hpp
+    #${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/Timer.hpp
 
     ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/scummvm/dosbox/dosbox.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/scummvm/mame/mame.hpp
@@ -161,6 +162,7 @@ target_sources(${LIB_NAME} PRIVATE
     # --- #
     ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/OPL.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/Config.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/Timer.cpp
 
     ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/scummvm/dosbox/dosbox.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/hardware/opl/scummvm/dosbox/dbopl/dbopl.cpp
